refact: add the `useEditor` hook
